Add gizmos/utilities.py, remove unnecessary late import
