+++ gizmos/diameter.py
@@ -8,9 +8,8 @@ from ..utilities.view import get_scale_from_pos
 from .base import (
     ConstraintGenericGGT,
     ConstraintGizmoGeneric,
-    draw_arrow_shape,
-    get_arrow_size,
 )
+from .utilities import draw_arrow_shape, get_arrow_size
 
 
 class VIEW3D_GGT_slvs_diameter(GizmoGroup, ConstraintGenericGGT):
